IDA types cleanup

+++ game/source/networking/delivery/network_link.hpp
@@ -7,11 +7,11 @@ struct c_network_out_of_band_consumer;
 struct c_network_link
 {
 	bool m_initialized;
-	int __unknown4;
-	int __unknown8;
+	long __unknown4;
+	long __unknown8;
 	transport_endpoint* m_endpoint;
 	c_network_out_of_band_consumer* m_out_of_band;
-	int __unknown14;
+	long __unknown14;
 	c_network_time_statistics m_time_statistics[4];
 };
 static_assert(sizeof(c_network_link) == 0x378);
